@@ -16,17 +16,17 @@
     "lint:types": "tsc",
     "lint:eslint": "eslint packages",
     "lint:all": "yarn lint:eslint && yarn lint:types",
-    "release:build": "node ./scripts/build/",
-    "release:test": "node ./scripts/test.js",
+    "release:build": "node -r @swc-node/register ./scripts/build/index.ts",
+    "release:test": "node -r @swc-node/register  ./scripts/test.ts",
     "version:next": "lerna version --exact --force-publish --no-private --preid next --create-release github --conventional-commits --conventional-prerelease --yes",
     "version:latest": "lerna version --exact --force-publish --no-private --create-release github --conventional-commits --yes",
-    "release:latest": "lerna publish from-package --contents build --dist-tag latest --yes",
-    "release:next": "lerna publish from-package --contents build --canary --preid next --pre-dist-tag next --yes",
+    "release:latest": "lerna publish from-package --dist-tag latest --yes",
+    "release:next": "lerna publish from-package --canary --preid next --pre-dist-tag next --yes",
     "build:docs": "cd docs && ./vercel-build.sh",
     "size-limit": "size-limit"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=14.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
@@ -37,11 +37,16 @@
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
     "@size-limit/preset-small-lib": "^4.10.1",
+    "@swc/jest": "^0.2.24",
+    "@swc/core": "^1.3.26",
+    "swc-node": "^1.0.0",
     "@testing-library/react": "^11.0.4",
     "@types/babel-types": "^7.0.9",
     "@types/jest": "^26.0.14",
+    "@types/node": "14.14.31",
     "@types/ramda": "^0.27.23",
     "@types/react": "^16.9.51",
+    "@types/rimraf": "^3.0.2",
     "@typescript-eslint/eslint-plugin": "^4.4.0",
     "@typescript-eslint/parser": "^4.4.0",
     "babel-code-frame": "^6.26.0",
@@ -76,7 +81,6 @@
     "node-emoji": "^1.10.0",
     "npm-cli-login": "^0.1.1",
     "ora": "^5.1.0",
-    "pkg-up": "^3.1.0",
     "prettier": "^2.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -94,7 +98,6 @@
     "set-tz": "^0.2.0",
     "size-limit": "^4.10.1",
     "strip-ansi": "^6.0.0",
-    "targz": "^1.0.1",
     "tmp": "^0.2.1",
     "ts-jest": "^26.4.1",
     "typescript": "^4.0.3",
