@@ -1,31 +1,43 @@
 import fs from "fs"
+import nodePath from "path"
 import pkgUp from "pkg-up"
+import cliPackageJson from "../../package.json";
 
 export const projectType = {
   CRA: "CRA",
   REACT: "REACT",
 }
 
-function getPackageJson() {
-  const packageJsonPath = pkgUp.sync()
+type PackageJson = {
+  name: string
+  dependencies: Record<string, string>
+  devDependencies: Record<string, string>
+}
+
+function getPackageJson(cwd?: string): {path: string, content: PackageJson} {
+  const packageJsonPath = pkgUp.sync({
+    cwd
+  })
 
   try {
     const json = fs.readFileSync(packageJsonPath, "utf8")
-    return JSON.parse(json)
+    return {path: packageJsonPath, content: JSON.parse(json)}
   } catch (e) {
     console.error(e)
     return null
   }
 }
 
-function hasDependency(pkg, name) {
+
+
+function hasDependency(pkg: PackageJson, name: string) {
   return (
     (pkg.dependencies && pkg.dependencies[name]) ||
     (pkg.devDependencies && pkg.devDependencies[name])
   )
 }
 
-function detectFramework(pkg) {
+function detectFramework(pkg: PackageJson) {
   if (hasDependency(pkg, "react-scripts")) {
     return projectType.CRA
   }
@@ -38,8 +50,13 @@ function detectFramework(pkg) {
 }
 
 export function detect() {
-  const pkg = getPackageJson()
+  let pkg = getPackageJson()
+
   if (!pkg) return null
 
-  return detectFramework(pkg)
+  if (pkg.content.name === cliPackageJson.name) {
+    pkg = getPackageJson(nodePath.dirname(pkg.path))
+  }
+
+  return detectFramework(pkg.content)
 }
